Fix index name
Previously, we did search start.spring.io stats using the initializr
index name. Such index name is actually an alias to the index of the
current year. This worked so far as it matched the query (query for
year 2020). Now that we've switched to 2021, our alias points to
initializr-2021 that doesn't have any 2020-related data.

This commit fixes the request to use another alias that encompass all
the yearly indexes.